# Conflicts:
#	src/main/java/com/epam/hotel/repositories/ReservationRepository.java
#	src/main/java/com/epam/hotel/services/ReservationService.java
#	src/main/java/com/epam/hotel/services/implementations/ReservationServiceImpl.java

+++ src/main/java/com/epam/hotel/services/implementations/UserServiceImpl.java
@@ -1,12 +1,17 @@
 package com.epam.hotel.services.implementations;
 
 import com.epam.hotel.domains.User;
+import com.epam.hotel.domains.enums.BlockStatus;
 import com.epam.hotel.domains.enums.Permission;
 import com.epam.hotel.repositories.UserRepository;
 import com.epam.hotel.services.UserService;
 import com.epam.hotel.utils.Encoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -36,6 +41,7 @@ public class UserServiceImpl implements UserService {
     public User save(User user) {
         if (user.getId() == 0) {
             user.setPermission(Permission.USER);
+            user.setBlock(BlockStatus.UNBLOCKED);
             user.setPassword(Encoder.encode(user.getPassword()));
         }
         return userRepository.save(user);
@@ -43,12 +49,34 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.get().getPermission() == Permission.USER){
+            userRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("You cant delete ADMIN");
+        }
     }
 
     @Override
     public User findByLogin(String login) {
         return userRepository.findByLogin(login);
+    }
+
+    @Override
+    public Page<User> getAllUsersPaged(int page, int size) { return userRepository.findAll(PageRequest.of(page - 1, size, Sort.Direction.ASC, "id")); }
+
+    @Override
+    public void changeUserBlockForId(Long id) {
+        User user = userRepository.getOne(id);
+        switch (user.getBlock()) {
+            case BLOCKED:
+                user.setBlock(BlockStatus.UNBLOCKED);
+                break;
+            case UNBLOCKED:
+                user.setBlock(BlockStatus.BLOCKED);
+                break;
+        }
+        userRepository.save(user);
     }
 
 }
